refactor(writeImage,Mesh): useCompression argument comes last
BREAKING_CHANGE: The `useCompression` argument now comes last in the functions that write images and meshes to array buffers and files

+++ src/io/writeArrayBuffer.ts
@@ -10,19 +10,23 @@ import Mesh from '../core/Mesh.js'
 
 import WriteArrayBufferResult from './WriteArrayBufferResult.js'
 
-async function writeArrayBuffer (webWorker: Worker | null, useCompression: boolean, imageOrMesh: Image | Mesh, fileName: string, mimeType: string): Promise<WriteArrayBufferResult> {
+async function writeArrayBuffer (webWorker: Worker | null, imageOrMesh: Image | Mesh, fileName: string, mimeType: string = '', useCompression: boolean = false): Promise<WriteArrayBufferResult> {
+  if (typeof imageOrMesh === 'boolean') {
+    throw new Error('useCompression is now the argument position in itk-wasm')
+  }
+
   const extension = getFileExtension(fileName)
   const isMesh = !!extensionToMeshIO.has(extension) || !!mimeToMeshIO.has(mimeType)
   if (isMesh) {
-    return await writeMeshArrayBuffer(webWorker, { useCompression }, imageOrMesh as Mesh, fileName, mimeType)
+    return await writeMeshArrayBuffer(webWorker, imageOrMesh as Mesh, fileName, mimeType, { useCompression })
       .catch(async function () {
         if (webWorker != null) {
           webWorker.terminate()
         }
-        return await writeImageArrayBuffer(null, useCompression, imageOrMesh as Image, fileName, mimeType)
+        return await writeImageArrayBuffer(null, imageOrMesh as Image, fileName, mimeType, useCompression)
       })
   } else {
-    return await writeImageArrayBuffer(webWorker, useCompression, imageOrMesh as Image, fileName, mimeType)
+    return await writeImageArrayBuffer(webWorker, imageOrMesh as Image, fileName, mimeType, useCompression)
   }
 }
